requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
passlib==1.7.4
python-multipart==0.0.17
python-jose==3.3.0
cryptography==43.0.3